\name{GBSA}
\alias{GBSA}
\title{
Averaging the called sites into customised bins.
}
\description{
Given intermediated files generated by format_grand.p or read.trio, this function bin the genome to consecutive bins and average gentoype calls across differnt bin
}
\usage{
GBSA(cutoffLevel,pedigreeTable,matchingNames,bin.size,vcf.file.f2,pathout,py.script)

}
\arguments{
    \item{cutoffLevel}{ The number of markers in each bin. Below this cutoff this bin is marked as NA,otherwise average is not used as calls}
    \item{pedigreeTable}{a data.frame with 7 columns with names as id.f2, id.f2.ma, id.f2.fa, fa.h, ma.h, ma.l, fa.l. Those are F2 id, corresponding mother id,father id and grand parent ids, fa.h,ma.h must come from one line and ma.l, fa.l. from another divergent line}
    \item{matchingNames}{a tab delimited file,documenting the chrmsome name length and contig names, see examples:https://github.com/yanjunzan/F2_re_seq/blob/master/data/chr_id.match.txt  only columns: RefSeq,INSDC,Type,Name,Size(Mb) are necessary }
    \item{bin.size}{ how large the size of the bin should be}
    \item{vcf.file.f2}{input vcf file}
    \item{pathout}{ path to the directory where intermediate files are stored}
}
\details{
}
\value{

return a data.table object with genotypes of each bin in column and individuals in each row.

}
\references{
Yanjun Zan, Thibaut Payen, Leif Andersson, Paul B. Siegel and Örjan Carlborg;Whole-genome QTL mapping in experimental pedigrees from outbred founders utilizing low coverage individual based sequencing 2018
}
\author{
Yanjun Zan, Thibaut Payen
}
\note{
}

\examples{

}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
